extend data_toolkit/nse module functionalities
